Move normalize under nze
