+++ dhall/src/semantics/nze/mod.rs
@@ -1,7 +1,9 @@
 pub mod env;
+pub mod normalize;
 pub mod value;
 pub mod var;
 pub mod visitor;
 pub(crate) use env::*;
+pub(crate) use normalize::*;
 pub(crate) use value::*;
 pub(crate) use var::*;
